Merge OCR file handling into Lasalle Branch (#29)
* OCR algorithm functions implemented

* File updates from git stash

* Updating default parameters for image processing

* Hooking up OCRMatching to the API

Updating OCR functions to take in upload files and adding API routing for the OCR. Needs testing

* Update requirements.txt

* OCR now properly return results (Dummy database only)

Changed Dockerfile and requirements for compatability with paddle. Issue was caused by depreceated function in older model of paddleX. Edited some code in the backend to allow algorithm to work with uploadfiles properly (bugpatch). Currently only hooked up to a dummy database. Will need to hook up to actual Session db

* Updating documentation and cleaning up code

* removing testing files

* Update Dockerfile

Adding system dependencies and RUN commands for ClamAV to the Dockerfile

* Implemented File scanning and cleaning

I leveraged clamAV in tandem with Pillow and shutil to scan, clean and quarantine all files uploaded to the backend. This should help protect the codebase from any malicious files that might get uploaded to the OCR

* Documenting New Code