Improved result filtering for OCR
Reject fuzzy results that have an = or > distance than the number of characters in the string. For example words like "ask" would be matched with "iii" with only a distance of three. This ranks high despite having nothing to do with the fuzzy matched result. This aims to avoid unlikely results being returned to the user